feat: merge nems functions
