--- application/CohortManager/src/Functions/DemographicServices/NEMSUnSubscription/NEMSUnSubscriptionConfig.cs
@@ -1,10 +1,0 @@
-namespace NHS.Screening.NEMSUnSubscription;
-
-using System.ComponentModel.DataAnnotations;
-
-public class NEMSUnSubscriptionConfig
-{
-    [Required]
-    public string NhsNumber { get; set; }
-    public string NemsDeleteEndpoint { get; set; }
-}
