--- application/CohortManager/src/Functions/DemographicServices/NEMSUnSubscription/host.json
@@ -1,12 +1,0 @@
-{
-  "version": "2.0",
-  "functionTimeout": "01:00:00",
-  "logging": {
-    "applicationInsights": {
-      "samplingExcludedTypes": "Request",
-      "samplingSettings": {
-        "isEnabled": true
-      }
-    }
-  }
-}
