+++ application/CohortManager/src/Functions/Functions.sln
@@ -137,13 +137,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "ServiceNowIntegrationServic
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "DemographicServices", "DemographicServices", "{0DDB5B2A-AD97-BFD6-A081-43D49FE0B20F}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NEMSSubscribe", "DemographicServices\NEMSSubscribe\NEMSSubscribe.csproj", "{800E7765-EC47-4343-A52C-16DF435B24D3}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DemographicDurableFunction", "DemographicServices\DemographicDurableFunction\DemographicDurableFunction.csproj", "{575CC2D2-BB5E-4CC2-AF65-E437F5F0263B}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DemographicDataManagementFunction", "DemographicServices\DemographicDataManagementFunction\DemographicDataManagementFunction.csproj", "{3E4AAF8E-EB14-46C2-9DF8-2A5734E8F2CD}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NEMSUnSubscription", "DemographicServices\NEMSUnSubscription\NEMSUnSubscription.csproj", "{21CEB638-C25E-4ABB-84BE-659819A8C9F7}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RetrievePDSDemographic", "DemographicServices\RetrievePDSDemographic\RetrievePDSDemographic.csproj", "{C0728070-DB44-492D-BDCD-B65FE5FC286A}"
 EndProject
@@ -262,6 +258,8 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ReceiveServiceNowMessage", "ServiceNowIntegrationService\ReceiveServiceNowMessage\ReceiveServiceNowMessage.csproj", "{0F5234AE-1991-491D-A7E2-0921EF90B7BA}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SendServiceNowMessageTests", "..\..\..\..\tests\UnitTests\SendServiceNowMessageTests\SendServiceNowMessageTests.csproj", "{10EADCB8-DF07-495F-8EBA-5A399962956D}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ManageNemsSubscription", "DemographicServices\ManageNemsSubscription\ManageNemsSubscription.csproj", "{800E7765-EC47-4343-A52C-16DF435B24D3}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -837,18 +835,6 @@ Global
 		{6DD9FFBC-738A-4B07-BCD6-7435B97701B2}.Release|x64.Build.0 = Release|Any CPU
 		{6DD9FFBC-738A-4B07-BCD6-7435B97701B2}.Release|x86.ActiveCfg = Release|Any CPU
 		{6DD9FFBC-738A-4B07-BCD6-7435B97701B2}.Release|x86.Build.0 = Release|Any CPU
-		{800E7765-EC47-4343-A52C-16DF435B24D3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{800E7765-EC47-4343-A52C-16DF435B24D3}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{800E7765-EC47-4343-A52C-16DF435B24D3}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{800E7765-EC47-4343-A52C-16DF435B24D3}.Debug|x64.Build.0 = Debug|Any CPU
-		{800E7765-EC47-4343-A52C-16DF435B24D3}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{800E7765-EC47-4343-A52C-16DF435B24D3}.Debug|x86.Build.0 = Debug|Any CPU
-		{800E7765-EC47-4343-A52C-16DF435B24D3}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{800E7765-EC47-4343-A52C-16DF435B24D3}.Release|Any CPU.Build.0 = Release|Any CPU
-		{800E7765-EC47-4343-A52C-16DF435B24D3}.Release|x64.ActiveCfg = Release|Any CPU
-		{800E7765-EC47-4343-A52C-16DF435B24D3}.Release|x64.Build.0 = Release|Any CPU
-		{800E7765-EC47-4343-A52C-16DF435B24D3}.Release|x86.ActiveCfg = Release|Any CPU
-		{800E7765-EC47-4343-A52C-16DF435B24D3}.Release|x86.Build.0 = Release|Any CPU
 		{575CC2D2-BB5E-4CC2-AF65-E437F5F0263B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{575CC2D2-BB5E-4CC2-AF65-E437F5F0263B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{575CC2D2-BB5E-4CC2-AF65-E437F5F0263B}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -873,18 +859,6 @@ Global
 		{3E4AAF8E-EB14-46C2-9DF8-2A5734E8F2CD}.Release|x64.Build.0 = Release|Any CPU
 		{3E4AAF8E-EB14-46C2-9DF8-2A5734E8F2CD}.Release|x86.ActiveCfg = Release|Any CPU
 		{3E4AAF8E-EB14-46C2-9DF8-2A5734E8F2CD}.Release|x86.Build.0 = Release|Any CPU
-		{21CEB638-C25E-4ABB-84BE-659819A8C9F7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{21CEB638-C25E-4ABB-84BE-659819A8C9F7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{21CEB638-C25E-4ABB-84BE-659819A8C9F7}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{21CEB638-C25E-4ABB-84BE-659819A8C9F7}.Debug|x64.Build.0 = Debug|Any CPU
-		{21CEB638-C25E-4ABB-84BE-659819A8C9F7}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{21CEB638-C25E-4ABB-84BE-659819A8C9F7}.Debug|x86.Build.0 = Debug|Any CPU
-		{21CEB638-C25E-4ABB-84BE-659819A8C9F7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{21CEB638-C25E-4ABB-84BE-659819A8C9F7}.Release|Any CPU.Build.0 = Release|Any CPU
-		{21CEB638-C25E-4ABB-84BE-659819A8C9F7}.Release|x64.ActiveCfg = Release|Any CPU
-		{21CEB638-C25E-4ABB-84BE-659819A8C9F7}.Release|x64.Build.0 = Release|Any CPU
-		{21CEB638-C25E-4ABB-84BE-659819A8C9F7}.Release|x86.ActiveCfg = Release|Any CPU
-		{21CEB638-C25E-4ABB-84BE-659819A8C9F7}.Release|x86.Build.0 = Release|Any CPU
 		{C0728070-DB44-492D-BDCD-B65FE5FC286A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{C0728070-DB44-492D-BDCD-B65FE5FC286A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C0728070-DB44-492D-BDCD-B65FE5FC286A}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -1593,6 +1567,18 @@ Global
 		{10EADCB8-DF07-495F-8EBA-5A399962956D}.Release|x64.Build.0 = Release|Any CPU
 		{10EADCB8-DF07-495F-8EBA-5A399962956D}.Release|x86.ActiveCfg = Release|Any CPU
 		{10EADCB8-DF07-495F-8EBA-5A399962956D}.Release|x86.Build.0 = Release|Any CPU
+		{800E7765-EC47-4343-A52C-16DF435B24D3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{800E7765-EC47-4343-A52C-16DF435B24D3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{800E7765-EC47-4343-A52C-16DF435B24D3}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{800E7765-EC47-4343-A52C-16DF435B24D3}.Debug|x64.Build.0 = Debug|Any CPU
+		{800E7765-EC47-4343-A52C-16DF435B24D3}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{800E7765-EC47-4343-A52C-16DF435B24D3}.Debug|x86.Build.0 = Debug|Any CPU
+		{800E7765-EC47-4343-A52C-16DF435B24D3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{800E7765-EC47-4343-A52C-16DF435B24D3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{800E7765-EC47-4343-A52C-16DF435B24D3}.Release|x64.ActiveCfg = Release|Any CPU
+		{800E7765-EC47-4343-A52C-16DF435B24D3}.Release|x64.Build.0 = Release|Any CPU
+		{800E7765-EC47-4343-A52C-16DF435B24D3}.Release|x86.ActiveCfg = Release|Any CPU
+		{800E7765-EC47-4343-A52C-16DF435B24D3}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -1655,15 +1641,14 @@ Global
 		{A14C561A-04DA-480E-8082-A92B604E275A} = {006A9B93-41E9-9D6B-03D8-80B6AB48B9E5}
 		{622371A9-8189-4E20-BD9A-0A3508073E36} = {A14C561A-04DA-480E-8082-A92B604E275A}
 		{6DD9FFBC-738A-4B07-BCD6-7435B97701B2} = {055277D8-0B3B-A388-B9E2-0CA913C660FF}
-		{800E7765-EC47-4343-A52C-16DF435B24D3} = {0DDB5B2A-AD97-BFD6-A081-43D49FE0B20F}
 		{575CC2D2-BB5E-4CC2-AF65-E437F5F0263B} = {0DDB5B2A-AD97-BFD6-A081-43D49FE0B20F}
 		{3E4AAF8E-EB14-46C2-9DF8-2A5734E8F2CD} = {0DDB5B2A-AD97-BFD6-A081-43D49FE0B20F}
-		{21CEB638-C25E-4ABB-84BE-659819A8C9F7} = {0DDB5B2A-AD97-BFD6-A081-43D49FE0B20F}
 		{C0728070-DB44-492D-BDCD-B65FE5FC286A} = {0DDB5B2A-AD97-BFD6-A081-43D49FE0B20F}
 		{5F0DAA07-D15F-460B-9FDA-B10AABD144C6} = {055277D8-0B3B-A388-B9E2-0CA913C660FF}
 		{68293DEC-C32E-4CD0-ADD0-EDF273017584} = {72FBAE8E-5F64-9545-D6D4-8F59D0E91151}
 		{43AF1482-6B3E-41C9-BF92-531EB00FEA20} = {4F52FD11-658E-A102-6CD3-7D7C16FFA15B}
 		{53D67486-6264-4554-BD21-F47811F16FEC} = {ADBDFBE0-1F6C-9F07-943C-8060402FBCEA}
 		{0F5234AE-1991-491D-A7E2-0921EF90B7BA} = {ADBDFBE0-1F6C-9F07-943C-8060402FBCEA}
+		{800E7765-EC47-4343-A52C-16DF435B24D3} = {0DDB5B2A-AD97-BFD6-A081-43D49FE0B20F}
 	EndGlobalSection
 EndGlobal
